services/risk-assessment/pyproject.toml
[project]
name = "risk-assessment"
version = "0.1.0"
description = "Dynamic Premium Calculation service for Myrva"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apache-airflow>=3.1.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "psycopg2>=2.9.11",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.48",
    "sqlalchemy-timescaledb>=0.4.1",
    "uvicorn>=0.42.0",
    "xgboost>=3.2.0",
]
